DNS: resolving: move all the resolving functions in another modules
